Revert "Squash merging from upstream's main branch (20260603)"
This reverts commit 5dd9e9d5f061a628847dde65227910dbcaa10101.

+++ dom/media/webaudio/AudioListener.h
@@ -1,3 +1,5 @@
+/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
+/* vim:set ts=2 sw=2 sts=2 et cindent: */
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -56,9 +58,9 @@ class AudioListener final : public nsWrapperCache {
   JSObject* WrapObject(JSContext* aCx,
                        JS::Handle<JSObject*> aGivenProto) override;
 
-  void SetPosition(float aX, float aY, float aZ);
-  void SetOrientation(float aX, float aY, float aZ, float aXUp, float aYUp,
-                      float aZUp);
+  void SetPosition(double aX, double aY, double aZ);
+  void SetOrientation(double aX, double aY, double aZ, double aXUp, double aYUp,
+                      double aZUp);
 
   AudioListenerEngine* Engine() { return mEngine.get(); }
