+++ dom/midi/MIDIPortParent.cpp
@@ -1,3 +1,5 @@
+/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
+/* vim:set ts=2 sw=2 sts=2 et cindent: */
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -89,7 +91,9 @@ bool MIDIPortParent::SendUpdateStatus(
     }
     mMessageQueue.Clear();
   }
-  return PMIDIPortParent::SendUpdateStatus(mDeviceState, mConnectionState);
+  return PMIDIPortParent::SendUpdateStatus(
+      static_cast<uint32_t>(mDeviceState),
+      static_cast<uint32_t>(mConnectionState));
 }
 
 MIDIPortParent::MIDIPortParent(const MIDIPortInfo& aPortInfo,
