+++ gfx/angle/checkout/src/libANGLE/renderer/d3d/VertexDataManager.h
@@ -112,7 +112,7 @@ class VertexDataManager : angle::NonCopyable
                                       GLint start,
                                       size_t count,
                                       GLsizei instances,
-                                      uint64_t baseInstance);
+                                      GLuint baseInstance);
 
     // Promote static usage of dynamic buffers.
     static void PromoteDynamicAttribs(const gl::Context *context,
@@ -142,14 +142,14 @@ class VertexDataManager : angle::NonCopyable
                                         GLint start,
                                         size_t count,
                                         GLsizei instances,
-                                        uint64_t baseInstance);
+                                        GLuint baseInstance);
 
     angle::Result storeDynamicAttrib(const gl::Context *context,
                                      TranslatedAttribute *translated,
                                      GLint start,
                                      size_t count,
                                      GLsizei instances,
-                                     uint64_t baseInstance);
+                                     GLuint baseInstance);
 
     BufferFactoryD3D *const mFactory;
