+++ gfx/angle/checkout/src/libANGLE/renderer/d3d/d3d9/Renderer9.h
@@ -368,7 +368,7 @@ class Renderer9 : public RendererD3D
                                          const gl::VertexBinding &binding,
                                          size_t count,
                                          GLsizei instances,
-                                         uint64_t baseInstance,
+                                         GLuint baseInstance,
                                          unsigned int *bytesRequiredOut) const override;
 
     angle::Result copyToRenderTarget(const gl::Context *context,
