+++ gfx/angle/checkout/src/libANGLE/renderer/d3d/d3d9/VertexBuffer9.h
@@ -30,7 +30,7 @@ class VertexBuffer9 : public VertexBuffer
                                         const gl::VertexAttribute &attrib,
                                         const gl::VertexBinding &binding,
                                         gl::VertexAttribType currentValueType,
-                                        size_t start,
+                                        GLint start,
                                         size_t count,
                                         GLsizei instances,
                                         unsigned int offset,
